Remove `SHUTDOWNABLE` signature
This follows the clarification that `V1_LWT.close` follows the behaviour
of the Mirage tcp/ip stack i.e. it signals a "shutdown" to the remote
(like sending an `Eof`) while leaving the resources allocated and
the connection up. Implementations which associate extra resources with
a connection (such as a file descriptor or socket buffers) need provide
some other cleanup method (e.g. a `disconnect` function)

Fixes #16

Signed-off-by: David Scott <dave@recoil.org>